Expand the full path and escape special characters.
If this is run in a directory with any special characters, the expanded
path will fail slightly randomly. Expanding was previously fine, but
this makes it clearer what's happening.

Additionally, this switches to expanding the path relative to the
current file, as in certain circumstances it's possible that RSpec has
switched the working directory to `spec/tmp` (which it does for
acceptance tests).

`shellescape` comes from `Shellwords` in the standard library, which is
already included.