Switch to use the Boxes Metadata API for Chef.
Recently, Chef changed the way that packages can be discovered. This
broke the original script.

Since, Chef has released a library (`Mixlib::Install`) which provides
the ability to look up versions of most of their projects. As that's in
Ruby (and is quite big) that's been implemented as an API external to
this.

Now, we can fetch version information from the Boxes Metadata API, fetch
and install the package.